Move interact to _container_io_manager
